+++ resources/lang/cs/gmuhk.php
@@ -1,0 +1,12 @@
+<?php
+
+return [
+    'work_types' => [
+        'F' => 'fotografie',
+        'G' => 'grafika',
+        'K' => 'kresba',
+        'O' => 'obraz',
+        'P' => 'plastika',
+        'VA' => 'video art',
+    ],
+];